Turn chart subtypes into renderer functions
In order to make charts extensible, define only a single generic type
"d3charts.Chart" in JS. Turn the concrete chart implementations into
renderer functions that return an object with a layout() and a render()
function.

Provide the renderer function as a property for now, it might later
become either JS code or a reference.

Instead of keeping a copy of the default config values in the JS code,
always render it to the client.