Pass widget reference to render function on creation
The renderer needs access to the widget to notify listeners.

Since a renderer instance is always bound to a widget instance, it's
sufficient to provide the widget reference to the renderer once upon
creation instead of a parameter to the render method that is must be
passed down the call chain.