Revise resource loading
To make charts extensible, don't concatenate the JavaScript files for
the three example chart types anymore, but let every chart subtype
load their corresponding resource files by themselves.

Provide a protected method require() that ensures that a given URL is
loaded.

Use an application-scoped instance of ChartResources that maintains a
map of the resources that are already registered.