Load d3.js and nvd3 from a CDN
When the widget includes the required JS and CSS files for d3.js and
nvd3.js, extensions of this widget are bound to the version shipped
with the base widget.

Since the widget is generic enough to support different versions of
d3, load the library files from a CDN at runtime. Later we could make
the URL configurable.

Introduce an abstract base class for nv-based charts to keep the URLs
at a single place.